mini-projects/conversational-ai-assistant/requirements.txt
boto3==1.38.18
python-dotenv==1.1.0
pytest==8.3.5